Refactor sink helpers
Moved to controller/sinks package since they aren't generic.

Removed superfluous logging to eliminate context arg.

Get namespace from ObjectRef instead of separate parameter. The
ContainerSource reconciler always forces this to be the same as the
source namespace.

Move setting the sink argument to a reconciler function since early
returns are easier to follow than nested ifs.

Change the condition reason from "NotFound" to "Missing" when the
Source has no sink ref, since the user action required is different.
This should really be validated by the webhook but we don't have a
webhook yet.
